+++ docs/.vitepress/config.ts
@@ -37,6 +37,8 @@ export default defineConfig({
           { text: 'Collections', link: '/guide/collections' },
           { text: 'Server Response', link: '/guide/server-response' },
           { text: 'Writing Resources', link: '/guide/writing-resources' },
+          { text: 'Generating Resources', link: '/guide/generating-resources' },
+          { text: 'Configuration', link: '/guide/configuration' },
         ]
       },
       {